Adding mock json for active/completed jobs
