--- active-jobs.json
@@ -1,3 +1,0 @@
-[
-  {"requestName":"Test123","quality":true,"security":true,"backup":true,"endTime":1556663509},{"requestName":"Test123","quality":true,"security":true,"backup":true,"endTime":1556663509}
-]
