+++ urbService/mock-active-jobs.json
@@ -1,0 +1,74 @@
+[
+  {
+    "id": 1,
+    "requestName": "Test Job",
+    "quality": true,
+    "security": false,
+    "backup": true,
+    "endTime": 1558314061,
+    "allocation": "east1"
+  },
+  {
+    "id": 2,
+    "requestName": "Hunter ML",
+    "quality": false,
+    "security": true,
+    "backup": true,
+    "endTime": 1558746061,
+    "allocation": "north2"
+  },
+  {
+    "id": 3,
+    "requestName": "CERN Test",
+    "quality": true,
+    "security": true,
+    "backup": false,
+    "endTime": 1558946061,
+    "allocation": "west1"
+  },
+  {
+    "id": 4,
+    "requestName": "GENI Experiment1",
+    "quality": true,
+    "security": false,
+    "backup": false,
+    "endTime": 1558326061,
+    "allocation": "west2"
+  },
+  {
+    "id": 5,
+    "requestName": "Team 3 Job",
+    "quality": false,
+    "security": true,
+    "backup": false,
+    "endTime": 1558216061,
+    "allocation": "east2"
+  },
+  {
+    "id": 9,
+    "requestName": "Team 2 Job",
+    "quality": true,
+    "security": true,
+    "backup": false,
+    "endTime": 1559116061,
+    "allocation": "east1"
+  },
+  {
+    "id": 10,
+    "requestName": "Active Test 1",
+    "quality": true,
+    "security": true,
+    "backup": true,
+    "endTime": 1558336061,
+    "allocation": "north1"
+  },
+  {
+    "id": 11,
+    "requestName": "NASA Test 1",
+    "quality": false,
+    "security": true,
+    "backup": true,
+    "endTime": 1558616061,
+    "allocation": "west2"
+  }
+]
