+++ urbService/mock-completed-jobs.json
@@ -1,0 +1,38 @@
+[
+  {
+    "id": 6,
+    "requestName": "Completed Job",
+    "quality": false,
+    "security": true,
+    "backup": true,
+    "endTime": 1556712061,
+    "allocation": "east1"
+  },
+  {
+    "id": 7,
+    "requestName": "Finished EXP",
+    "quality": false,
+    "security": true,
+    "backup": true,
+    "endTime": 1553712061,
+    "allocation": "west1"
+  },
+  {
+    "id": 8,
+    "requestName": "Some test123",
+    "quality": true,
+    "security": false,
+    "backup": false,
+    "endTime": 1555112061,
+    "allocation": "north2"
+  },
+  {
+    "id": 12,
+    "requestName": "Expired Job 456",
+    "quality": true,
+    "security": false,
+    "backup": false,
+    "endTime": 1554112061,
+    "allocation": "north2"
+  }
+]
